+++ src/lib.rs
@@ -11,7 +11,6 @@ pub mod bar;
 pub mod block;
 pub mod config;
 mod error;
-pub mod logger;
 pub mod protocol;
 
 pub use error::Error;